Corregidos problemas de borrado de usuarios

+++ src/Controller/UsersController.php
@@ -21,6 +21,7 @@ class UsersController extends AppController
         session_start();
         $encontrado = false;
         $_SESSION['rol'] = "";
+        $_SESSION['user'] = "false";
 
         if ($this->request->is(['patch', 'post', 'put']))
             {
@@ -41,11 +42,12 @@ class UsersController extends AppController
                     if($user->passwor == $_SESSION['password'])       //En caso de coincidir la contraseña
                     {
                         $_SESSION['rol'] = $user->role; 
-                        $encontrado = true;  
+                        $encontrado = true; 
+                        $_SESSION['user'] = "true"; 
                     }
                 }
             }
-            if ($encontrado)
+            if ($_SESSION['user'] == "true")
             {
                 $users = $this->paginate($this->Users);     //Lista todos los usuarios   
                 $this->set(compact('users'));
@@ -55,12 +57,6 @@ class UsersController extends AppController
             {
                 $this->Flash->error(__('Nombre de usuario o contraseña incorrectos'));
             }
-        }
-
-        else 
-        {
-            
-            
         }
     }
 
@@ -73,12 +69,21 @@ class UsersController extends AppController
      */
     public function view($id = null)
     {
-        $user = $this->Users->get($id, [
+        session_start();
+        if($_SESSION['user'] == "true")
+        {
+            $user = $this->Users->get($id, [
             'contain' => []
         ]);
 
-        $this->set('user', $user);
-        $this->set('_serialize', ['user']);
+            $this->set('user', $user);
+            $this->set('_serialize', ['user']);
+        }
+        else
+        {
+            return $this->redirect(['action' => 'index']);
+        }
+        
     }
 
     /**
@@ -88,18 +93,28 @@ class UsersController extends AppController
      */
     public function add()
     {
-        $user = $this->Users->newEntity();
-        if ($this->request->is('post')) {
-            $user = $this->Users->patchEntity($user, $this->request->getData());
-            if ($this->Users->save($user)) {
-                $this->Flash->success(__('The user has been saved.'));
+        session_start();
+        if($_SESSION['user'] == "true")
+        {
+            $user = $this->Users->newEntity();
+            if ($this->request->is('post')) 
+            {
+                $user = $this->Users->patchEntity($user, $this->request->getData());
+                if ($this->Users->save($user)) {
+                    $this->Flash->success(__('The user has been saved.'));
 
-                return $this->redirect(['action' => 'index']);
+                    return $this->redirect(['action' => 'index']);
             }
             $this->Flash->error(__('The user could not be saved. Please, try again.'));
         }
-        $this->set(compact('user'));
-        $this->set('_serialize', ['user']);
+            $this->set(compact('user'));
+            $this->set('_serialize', ['user']);
+        }
+        else
+        {
+            return $this->redirect(['action' => 'index']);
+        }
+        
     }
 
     /**
@@ -111,20 +126,29 @@ class UsersController extends AppController
      */
     public function edit($id = null)
     {
-        $user = $this->Users->get($id, [
+        session_start();
+        if($_SESSION['user'] == "true")
+        {
+            $user = $this->Users->get($id, [
             'contain' => []
         ]);
-        if ($this->request->is(['patch', 'post', 'put'])) {
-            $user = $this->Users->patchEntity($user, $this->request->getData());
-            if ($this->Users->save($user)) {
-                $this->Flash->success(__('The user has been saved.'));
+            if ($this->request->is(['patch', 'post', 'put'])) {
+                $user = $this->Users->patchEntity($user, $this->request->getData());
+                if ($this->Users->save($user)) {
+                    $this->Flash->success(__('The user has been saved.'));
 
-                return $this->redirect(['action' => 'index']);
+                    return $this->redirect(['action' => 'index']);
+                }
+                $this->Flash->error(__('The user could not be saved. Please, try again.'));
             }
-            $this->Flash->error(__('The user could not be saved. Please, try again.'));
+            $this->set(compact('user'));
+            $this->set('_serialize', ['user']);
         }
-        $this->set(compact('user'));
-        $this->set('_serialize', ['user']);
+        else
+        {
+            return $this->redirect(['action' => 'index']);
+        }
+        
     }
 
     /**
@@ -136,15 +160,24 @@ class UsersController extends AppController
      */
     public function delete($id = null)
     {
-        $this->request->allowMethod(['post', 'delete']);
-        $user = $this->Users->get($id);
-        if ($this->Users->delete($user)) {
-            $this->Flash->success(__('The user has been deleted.'));
-        } else {
-            $this->Flash->error(__('The user could not be deleted. Please, try again.'));
-        }
+        session_start();
+        if($_SESSION['user'] == "true")
+        {
+            $this->request->allowMethod(['post', 'delete']);
+            $user = $this->Users->get($id);
+            if ($this->Users->delete($user)) {
+                $this->Flash->success(__('The user has been deleted.'));
+            } else {
+                $this->Flash->error(__('The user could not be deleted. Please, try again.'));
+            }
 
-        return $this->redirect(['action' => 'index']);
+            return $this->redirect(['action' => 'index']);
+        }
+        else
+        {
+            return $this->redirect(['action' => 'index']);
+        }
+        
     }
 
     public function logout()
@@ -153,6 +186,7 @@ class UsersController extends AppController
         unset($_SESSION['usuario']); 
         unset($_SESSION['password']);
         unset($_SESSION['rol']);
+        unset($_SESSION['user']);
         session_destroy();
         return $this->redirect(['action' => 'index']);
     }
